+++ common/src/width_helpers.rs
@@ -1,14 +1,14 @@
-pub fn signed_smallest_width_of_value(value: i64) -> u32 {
+pub fn signed_smallest_width_of_value(value: i128) -> u32 {
     let value = if value.is_negative() {
         value.abs() - 1
     } else {
         value.abs()
     };
 
-    unsigned_smallest_width_of_value(u64::try_from(value).unwrap()) + 1 // +1 becuase it's a signed value
+    unsigned_smallest_width_of_value(u128::try_from(value).unwrap()) + 1 // +1 becuase it's a signed value
 }
 
-pub fn unsigned_smallest_width_of_value(value: u64) -> u32 {
+pub fn unsigned_smallest_width_of_value(value: u128) -> u32 {
     u32::try_from((value + 1).next_power_of_two().ilog2()).unwrap()
 }
